DWR/Youtube/youtube_backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
crispy-bootstrap3==2022.1
crispy-bootstrap4==2022.1
cron-descriptor==1.2.35
cryptography==40.0.1
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
django-import-export==3.1.0
django-rest-passwordreset==1.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
et-xmlfile==1.1.0
eventlet==0.33.3
fabric==3.0.0
fake==0.8
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
httpie==3.2.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
invoke==2.0.0
kombu==5.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupPy==1.14
mdurl==0.1.2
msgpack-python==0.5.6
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
paramiko==3.1.0
Pillow==9.4.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==23.1.1
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
redis==2.10.6
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
websockets==11.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
zope.interface==6.0
